Update icons & add png variants
^

The image assets stored in the server's /static/icons endpoint are not up-to-date (Still had the very old icon 26, which was requested to be removed/updated a while ago).

This commit updates this endpoint to have the latest icons and png variants of said icons.

This commit also adds a .gif version of the loading icon, the guest icon, and other misc icons such as announcements and alerts.